Added a common base test for *InstTest
The base test drags eliminates the TestUtil and makes it so that implementing tests only need to extend it and add a "name" for the test and the Table3 data. This base class  also provides for a comment() which can be used as (1) a header for sections and (2) to help locate the line of data in the table that has some failure as the "last comment" is printed on failure of an assertion. Also note that by convention, test data should start on a zero line in the code so that the math remains simple to find the failing line since scalatest does output the line number of the table that failed.